rename `global_heap` -> `libc_heap`
This module only contains wrappers for malloc and realloc with
out-of-memory checks.
